+++ tools/pokerus-tests/optional/kotlin/Pair.java
@@ -1,0 +1,2 @@
+package kotlin;
+public final class Pair<A, B> {}