+++ tools/pokerus-tests/stubs/com/cobblemon/mod/common/pokemon/Pokemon.java
@@ -1,0 +1,7 @@
+package com.cobblemon.mod.common.pokemon;
+public final class Pokemon {
+    public boolean infected;
+    public boolean reject;
+    public boolean fail;
+    public int infectionCalls;
+}